dashboard/requirements.txt
altair==4.1.0
appdirs==1.4.4
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto3==1.13.19
botocore==1.16.19
branca==0.4.1
Brotli==1.0.7
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
dash==1.9.1
dash-auth==1.0.0
dash-core-components==1.8.1
dash-daq==0.4.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dnspython==1.16.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
Flask-SeaSurf==0.2.2
folium==0.10.1
future==0.18.2
gunicorn==20.0.4
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.1
packaging==20.4
pandas==1.0.1
pandas-datareader==0.8.1
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.5.4
PriceIndices==1.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2==2.8.5
pydeck==0.3.1
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.1
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
streamlit==0.60.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
virtualenv==20.0.21
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
